🚧 AR080K context: fix dashboard CI checks
