+++ .agentplane/tasks/202606011716-AR080K/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202606011716-AR080K/add-fast-context-knowledge-dashboard",
   "created_at": "2026-06-01T17:17:02.350Z",
-  "diffstat_sha256": "sha256:e28c7c98ef4cc4679ea1c00e139d6f46cd1802bdd80abb26295ccb5ca54aba4a",
+  "diffstat_sha256": "sha256:7fb56b65660b354b042899a36b070475e4d9ae005f6413179d564fb7be171fd7",
   "last_verified_at": "2026-06-01T17:33:37.538Z",
   "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
   "schema_version": 1,
